requirements.txt
requests
beautifulsoup4
lxml
psycopg[binary]
python-dateutil
